refactor(test): isolate suggestion status-feed coverage
Hive-Task-Slug: add-repository-aware-answer-suggestions-260826-528d
Hive-Fix-Pass: 02
Hive-Fix-Findings: 1
Hive-Triage-Bias: courageous
Hive-Reviewer-Sources: codex-ce-code-review
Hive-Fix-Phase: fix
